Handle database replication delay when adding todos
This change adds logic to ensure newly-added todos are in the
returned list. When the Mattermost server running the plugin has
database read replicas configured, it is possible that the plugin
would miss listing the new todo due to replication delay. This
plugin logic attempts to address the issue when it is detected.